testing: 코드 정리
